+++ mango-core/src/main/java/org/calrissian/mango/domain/TemporalIdentifiable.java
@@ -1,0 +1,4 @@
+package org.calrissian.mango.domain;
+
+public interface TemporalIdentifiable extends Temporal, Identifiable {
+}
