+++ mango-core/src/main/java/org/calrissian/mango/domain/TypedIdentifiable.java
@@ -1,0 +1,6 @@
+package org.calrissian.mango.domain;
+
+public interface TypedIdentifiable extends Identifiable {
+
+    String getType();
+}
