+++ mango-core/src/main/java/org/calrissian/mango/domain/event/BaseEvent.java
@@ -13,14 +13,16 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.calrissian.mango.domain;
+package org.calrissian.mango.domain.event;
 
+
+import org.calrissian.mango.domain.BaseTupleStore;
 
 import static java.lang.System.currentTimeMillis;
 import static java.util.UUID.randomUUID;
 
 /**
- * Default implementation of {@link org.calrissian.mango.domain.Event}
+ * Default implementation of {@link Event}
  */
 public class BaseEvent extends BaseTupleStore implements Event {
