+++ mango-core/src/main/java/org/calrissian/mango/domain/ip/IPv4.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.calrissian.mango.domain;
+package org.calrissian.mango.domain.ip;
 
 
 import java.io.Serializable;
