+++ mango-core/src/main/java/org/calrissian/mango/types/SimpleTypeEncoders.java
@@ -16,7 +16,7 @@
 package org.calrissian.mango.types;
 
 
-import org.calrissian.mango.domain.IPv4;
+import org.calrissian.mango.domain.ip.IPv4;
 import org.calrissian.mango.types.encoders.*;
 import org.calrissian.mango.types.exception.TypeDecodingException;
