+++ mango-core/src/test/java/org/calrissian/mango/collect/TupleStoreIteratorTest.java
@@ -16,8 +16,8 @@
 package org.calrissian.mango.collect;
 
 
-import org.calrissian.mango.domain.BaseEntity;
-import org.calrissian.mango.domain.Entity;
+import org.calrissian.mango.domain.entity.BaseEntity;
+import org.calrissian.mango.domain.entity.Entity;
 import org.calrissian.mango.domain.Tuple;
 import org.junit.Before;
 import org.junit.Test;
