+++ mango-core/src/test/java/org/calrissian/mango/criteria/domain/criteria/HasNotCriteriaTest.java
@@ -15,8 +15,8 @@
  */
 package org.calrissian.mango.criteria.domain.criteria;
 
-import org.calrissian.mango.domain.BaseEntity;
-import org.calrissian.mango.domain.Entity;
+import org.calrissian.mango.domain.entity.BaseEntity;
+import org.calrissian.mango.domain.entity.Entity;
 import org.calrissian.mango.domain.Tuple;
 import org.junit.Test;
