+++ mango-core/src/test/java/org/calrissian/mango/types/SimpleTypeEncodersTest.java
@@ -16,7 +16,7 @@
 package org.calrissian.mango.types;
 
 
-import org.calrissian.mango.domain.IPv4;
+import org.calrissian.mango.domain.ip.IPv4;
 import org.calrissian.mango.types.exception.TypeDecodingException;
 import org.calrissian.mango.types.exception.TypeEncodingException;
 import org.junit.Test;
